Create SimulationManager for handling all updates to chart/calculations; this prevents unnecessary re-rendering of components such as ChartExplanation and Introduction
